requirements.txt
accelerate==1.10.1
ag2==0.9.10
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aistudio-sdk==0.3.8
alibabacloud-credentials==1.0.2
alibabacloud-credentials-api==1.0.0
alibabacloud-ecs20140526==7.2.3
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_util==0.3.13
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
APScheduler==3.11.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
azure-common==1.1.28
azure-core==1.35.1
azure-identity==1.25.1
azure-mgmt-compute==37.0.1
azure-mgmt-core==1.6.0
azure-mgmt-network==29.0.0
backoff==2.2.1
bce-python-sdk==0.9.46
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
borb==3.0.2
boto3==1.40.50
botocore==1.40.50
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli==1.1.0
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==6.2.0
certifi @ file:///home/task_175978570070155/conda-bld/certifi_1759786060979/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1758716105649/work
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cli_exit_tools==1.2.7
click==8.1.8
cloudpickle==3.1.1
cloudscraper==1.2.71
colorlog==6.9.0
contourpy==1.3.2
cryptography==44.0.3
cssselect==1.3.0
cycler==0.12.1
darabonba-core==1.0.3
dashscope==1.24.6
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
dotenv==0.9.9
easyocr==1.7.2
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evdev==1.9.2
exceptiongroup==1.3.0
fabric==3.2.2
Farama-Notifications==0.0.4
fastapi==0.119.0
fastdtw==0.3.4
ffmpy==0.6.3
filelock==3.20.0
Flask==3.0.3
fonttools==4.60.1
formulas==1.3.1
frozenlist==1.8.0
fsspec==2025.9.0
func_timeout==4.3.5
future==1.0.0
fuzz==0.1.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.43.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
groq==0.32.0
grpcio==1.75.1
grpcio-status==1.71.2
# Editable install with no version control (gui-agents==0.3.1)
-e /nvme/yangbowen/yangbowen/OSWorld/Agent-S
gymnasium==0.28.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
idna @ file:///croot/idna_1714398848350/work
ImageHash==4.3.2
imageio==2.37.0
imagesize==1.4.1
inflate64==1.0.3
invoke==2.2.1
isodate==0.7.2
itsdangerous==2.2.0
jax-jumpy==1.0.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
JSON-minify==0.3.0
json_repair==0.52.0
kiwisolver==1.4.9
lazy_loader==0.4
lib-detect-testenv==2.0.8
librosa==0.11.0
llvmlite==0.45.1
LnkParse3==1.5.3
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.5
mdurl==0.1.2
modelscope==1.30.0
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
msrest==0.7.1
multidict==6.7.0
multiprocess==0.70.18
multivolumefile==0.2.3
mutagen==1.47.0
networkx==3.4.2
ninja==1.13.0
numba==0.62.1
numexpr @ file:///croot/numexpr_1752521720723/work
numpy==1.24.4
numpy-financial==1.0.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
oauthlib==3.3.1
odfpy==1.4.1
openai==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.8.1.78
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.3
outcome==1.3.0.post0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.2.3
paramiko==4.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.5.0
playwright==1.50.0
pooch==1.8.2
prettytable==3.16.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
py-cpuinfo==9.0.0
py7zr==1.0.0
pyacoustid==1.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybcj==1.0.6
pyclipper==1.3.0.post6
pycparser @ file:///home/task_175749556993857/conda-bld/pycparser_1757496041120/work
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
PyDrive==1.3.1
pydub==0.25.1
pyee==12.1.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
PyMuPDF @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pymupdf_1749732479/work
PyNaCl==1.6.0
pynput==1.7.8
pyparsing==3.2.5
pypdf==6.1.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.11.0
pyppmd==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzstd==0.18.0
RapidFuzz==3.14.1
regex @ file:///croot/regex_1736540786412/work
requests @ file:///croot/requests_1756709366904/work
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.14.1
s3transfer==0.14.0
safehttpx==0.1.6
safetensors==0.6.2
schedula==1.5.64
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
selenium==4.36.0
semantic-version==2.10.0
sentry-sdk==2.41.0
shapely==2.1.2
shellingham==1.5.4
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
soxr @ file:///home/conda/feedstock_root/build_artifacts/soxr-python_1757596842799/work
starlette==0.48.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken @ file:///croot/tiktoken_1748479152668/work
tldextract==5.3.0
together==1.5.26
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.13.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.35.2
trio==0.31.0
trio-websocket==0.12.2
triton==3.1.0
typer==0.15.4
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.37.0
wandb==0.22.2
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wrapt_timeout_decorator==1.5.1
wsproto==1.2.0
xlsxwriter==3.2.9
xmltodict==1.0.2
yarl==1.22.0
